+++ src/main/java/com/strangegrotto/wealthdraft/assetallocation/TargetAssetAllocation.java
@@ -2,10 +2,13 @@ package com.strangegrotto.wealthdraft.assetallocation;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
+import com.google.common.base.Preconditions;
 import com.strangegrotto.wealthdraft.WealthdraftImmutableStyle;
+import com.strangegrotto.wealthdraft.assetallocation.filters.AssetFilter;
 import org.immutables.value.Value;
 
 import java.math.BigDecimal;
+import java.util.Map;
 import java.util.Optional;
 
 @WealthdraftImmutableStyle
@@ -13,7 +16,7 @@ import java.util.Optional;
 @JsonDeserialize(as = ImmTargetAssetAllocation.class)
 public interface TargetAssetAllocation {
     @JsonProperty("numerator")
-    AssetFilter getNumeratorFilter();
+    String getNumeratorFilter();
 
     // Fraction meaning 0.7, rather than 70%
     @JsonProperty("fraction")
@@ -21,5 +24,5 @@ public interface TargetAssetAllocation {
 
     // If not present, the denominator is assumed to be the entire portfolio
     @JsonProperty("denominator")
-    Optional<ConjunctiveAssetTagFilter> getDenominatorOpt();
+    Optional<String> getDenominatorFilterOpt();
 }