+++ src/test/java/com/strangegrotto/wealthdraft/assetallocation/TargetAssetAllocationTest.java
@@ -4,12 +4,16 @@ import com.fasterxml.jackson.databind.type.CollectionType;
 import com.strangegrotto.wealthdraft.Main;
 import com.strangegrotto.wealthdraft.assets.definition.AssetDefinitions;
 import com.strangegrotto.wealthdraft.assets.definition.AssetDefinitionsTestFiles;
+import com.strangegrotto.wealthdraft.assets.definition.ExpectedExampleAssetDefinitions;
+import com.strangegrotto.wealthdraft.assets.definition.ImmAssetDefinitions;
 import org.junit.Assert;
 import org.junit.Test;
 
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
+import java.util.Set;
 
 public class TargetAssetAllocationTest {
     @Test
@@ -18,11 +22,24 @@ public class TargetAssetAllocationTest {
         Assert.assertEquals(targetAssetAllocations, ExpectedExampleTargetAssetAllocations.ASSET_ALLOCATIONS);
     }
 
+    @Test
+    public void testAllSetMathOperators() throws IOException {
+        var targetAssetAllocations = parseAssetAllocationsFile(TargetAssetAllocationsTestFiles.ALL_SET_MATH_OPERATORS);
+        var megaFilter = targetAssetAllocations.getFilters().get("megaFilter");
+        var filteredAssetIds = megaFilter.apply(
+                ExpectedExampleAssetDefinitions.EXPECTED_ASSETS,
+                ExpectedExampleAssetDefinitions.EXPECTED_ASSETS.keySet()
+        );
+        var expectedAssetIds = Set.of(
+                ExpectedExampleAssetDefinitions.BROKERAGE_ACCOUNT_ID,
+                ExpectedExampleAssetDefinitions.BITCOIN_HOLDING_ID
+        );
+        Assert.assertEquals(expectedAssetIds, filteredAssetIds);
+    }
 
-    private static List<TargetAssetAllocation> parseAssetAllocationsFile(TargetAssetAllocationsTestFiles testFile) throws IOException {
+    private static TargetAssetAllocations parseAssetAllocationsFile(TargetAssetAllocationsTestFiles testFile) throws IOException {
         var mapper = Main.getObjectMapper();
         var assetAllocationsUrl = testFile.getResource();
-        CollectionType assetAllocationsListType = mapper.getTypeFactory().constructCollectionType(ArrayList.class, TargetAssetAllocation.class);
-        return mapper.readValue(assetAllocationsUrl, assetAllocationsListType);
+        return mapper.readValue(assetAllocationsUrl, TargetAssetAllocations.class);
     }
 }